# api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "api/api.proto",
    "version": "0.0.0",
    "contact": {
      "name": "NoCloud",
      "url": "https://github.com/slntopp/nocloud",
      "email": "info@slnt-opp.xyz"
    }
  },
...
  "host": "localhost:8000",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json",
    "text/event-stream"
  ],
  "produces": [
    "application/json",
    "text/event-stream"
  ],
  "paths": {
...
    "/logging/trace/{requestor}": {
      "get": {
        "operationId": "EventsLoggingService_GetTrace",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/nocloudevents_loggingEvents"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "requestor",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "EventsLoggingService"
        ]
      }
    },
    "/logging/{uuid}": {
      "get": {
        "operationId": "EventsLoggingService_GetEvents",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/nocloudevents_loggingEvents"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "entity",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "scope",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "EventsLoggingService"
        ]
      }
    },
...
  },
  "definitions": {
...
    "events_loggingSnapshot": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "diff": {
          "type": "string"
        }
      }
    },
    "googlerpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
...
    "nocloudevents_loggingEvent": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "entity": {
          "type": "string"
        },
        "uuid": {
          "type": "string"
        },
        "scope": {
          "type": "string"
        },
        "action": {
          "type": "string"
        },
        "rc": {
          "type": "integer",
          "format": "int32"
        },
        "requestor": {
          "type": "string"
        },
        "ts": {
          "type": "string",
          "format": "int64"
        },
        "snapshot": {
          "$ref": "#/definitions/events_loggingSnapshot"
        }
      }
    },
    "nocloudevents_loggingEvents": {
      "type": "object",
      "properties": {
        "events": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/nocloudevents_loggingEvent"
          }
        },
        "count": {
          "type": "string",
          "format": "int64"
        }
      }
    },
...
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
...
  }
}